+++ ms-reservas/src/main/java/com/eventos/ms_reservas/exception/SolicitudNotFoundException.java
@@ -5,14 +5,14 @@ package com.eventos.ms_reservas.exception;
  */
 public class SolicitudNotFoundException extends RuntimeException {
 
-    private final Long id;
+    private final Integer id;
 
-    public SolicitudNotFoundException(Long id, String message) {
+    public SolicitudNotFoundException(Integer id, String message) {
         super(message);
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
